+++ docker/kms/container/start.sh
@@ -4,12 +4,20 @@ exec 2>&1
 set -e
 set -x
 
-tmkms init /root/.tmkms
+echo "Staring KMS..."
+EXECUTED_CHECK="/root/executed"
 
-# tmkms softsign keygen /root/.tmkms/secret_connection.key
-mv $COMMON_DIR/tmkms.toml /root/.tmkms/
+if [ -f "$EXECUTED_CHECK" ]; then
+  sleep 10 && tmkms start -v -c /root/.tmkms/tmkms.toml
+else
+  tmkms init /root/.tmkms
 
-cd /root/.tmkms/ && ls
-tmkms softsign import $COMMON_DIR/priv_validator_key.json /root/.tmkms/signing.key
+  # tmkms softsign keygen /root/.tmkms/secret_connection.key
+  mv $COMMON_DIR/tmkms.toml /root/.tmkms/
 
-sleep 10 && tmkms start -v -c /root/.tmkms/tmkms.toml
+  cd /root/.tmkms/ && ls
+  tmkms softsign import $COMMON_DIR/priv_validator_key.json /root/.tmkms/signing.key
+
+  touch $EXECUTED_CHECK
+  sleep 10 && tmkms start -v -c /root/.tmkms/tmkms.toml
+fi
