+++ docker/sentry/container/start.sh
@@ -6,11 +6,19 @@ set -x
 
 echo "Staring sentry..."
 SEKAID_HOME=$HOME/.simapp
-rm -rf $SEKAID_HOME
 
-sekaid init --chain-id=testing testing --home=$SEKAID_HOME
-cp $COMMON_DIR/genesis.json $SEKAID_HOME/config/genesis.json
-cp $COMMON_DIR/node_key.json $SEKAID_HOME/config/node_key.json
-cp $COMMON_DIR/config.toml $SEKAID_HOME/config/config.toml
+EXECUTED_CHECK="/root/executed"
 
-sekaid start --home=$SEKAID_HOME --rpc.laddr="tcp://0.0.0.0:26657"
+if [ -f "$EXECUTED_CHECK" ]; then
+  sekaid start --home=$SEKAID_HOME --rpc.laddr="tcp://0.0.0.0:26657"
+else
+  rm -rf $SEKAID_HOME
+
+  sekaid init --chain-id=testing testing --home=$SEKAID_HOME
+  cp $COMMON_DIR/genesis.json $SEKAID_HOME/config/genesis.json
+  cp $COMMON_DIR/node_key.json $SEKAID_HOME/config/node_key.json
+  cp $COMMON_DIR/config.toml $SEKAID_HOME/config/config.toml
+
+  touch $EXECUTED_CHECK
+  sekaid start --home=$SEKAID_HOME --rpc.laddr="tcp://0.0.0.0:26657"
+fi
